requirements/base.txt
alembic==1.6.5
click==8.0.1
dynaconf==3.1.4
decorator==4.4.1
Flask==2.0.1
Flask-JWT-Extended==4.2.1
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
tox==3.23.1
virtualenv==20.4.7
Werkzeug==2.0.1
wrapt==1.12.1
marshmallow-enum==1.5.1
marshmallow==3.10.0
requests==2.25.1
webargs==8.0.0
flask-cors==3.0.10
requests-toolbelt
python-json-logger
